KAS-2834 new components for accesslevel & confidentiality
