+++ app/components/utils/access-level-color-badge.hbs
@@ -1,0 +1,2 @@
+<Auk::ColorBadge @skin={{this.accessPillSkin}}/>
+<span>{{@accessLevel.label}}</span>
