+++ app/components/utils/confidentiality-pill.hbs
@@ -1,0 +1,5 @@
+{{#if @confidential}}
+  <Auk::Pill @skin="danger" @layout="icon-only" @icon="lock-closed"/>
+{{else}}
+  <Auk::Pill @skin="success" @layout="icon-only" @icon="lock-open"/>
+{{/if}}
