+++ app/components/utils/confidentiality-pill.js
@@ -1,0 +1,6 @@
+import Component from '@glimmer/component';
+
+
+export default class ConfidentialityPill extends Component {
+
+}
